The new Matlab Scripts
I upload the new Matlab scripts used for the creation of the set that
is the bases of the Neural Network. The only script that the user needs
to open and run is the script ‘A1_Parametrization.m’. Currently, that
script saves outputs and intermediate data at the same path as the
script is saved in. This is not ideal, as this folder is monitored by
GitHub. For that reason, I will need to change it slightly. Also, it
now creates a .csv of the set. Since we are switching for Python to
Matlab for the Neural Network, it makes sense to export the set as a
.mat file, or not export at all, but start the Network from the RAM.